requirements.txt
discord.py==2.3.2
aiohttp==3.9.1
python-dotenv==1.0.0
cryptography==41.0.7
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.23
alembic==1.13.1
redis==5.0.1
loguru==0.7.2
click==8.1.7
tenacity==8.2.3
pydantic==2.5.3
fastapi==0.108.0
uvicorn==0.25.0
psycopg2-binary==2.9.9